+++ src/pages/Asset/Asset.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Asset:React.FC = () => {
+  return (
+    <>
+      开发中喽.....
+    </>
+  )
+}
+
+export default Asset
