+++ src/pages/Login/Login.tsx
@@ -1,0 +1,15 @@
+/**
+ * @desc Login 登录页
+ * @date 2023/11/06
+ * @author 7_mu12
+ */
+import React from "react";
+//todo
+const Login: React.FC = () => {
+  return (
+    <>
+      Login
+    </>
+  )
+}
+export default Login
